Integrate with standard release pipeline
