--- .github/workflows/pro-release-develop.yml
@@ -1,47 +1,0 @@
-name: Budibase Release Staging (Pro)
-
-#  Temporary pipeline - eventualy this will be merged with the regular one
-
-on:
-  push:
-    branches:
-      - pro-develop
-    paths:
-      - ".aws/**"
-      - ".github/**"
-      - "charts/**"
-      - "packages/**"
-      - "scripts/**"
-      - "package.json"
-      - "yarn.lock"
-      - "package.json"
-      - "yarn.lock"
-
-env:
-  POSTHOG_TOKEN: ${{ secrets.POSTHOG_TOKEN }}
-  INTERCOM_TOKEN: ${{ secrets.INTERCOM_TOKEN }}
-  POSTHOG_URL: ${{ secrets.POSTHOG_URL }}
-
-jobs:
-  release:
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-node@v1
-        with:
-          node-version: 14.x
-      - run: yarn
-      - run: yarn bootstrap
-      - run: yarn lint
-      - run: yarn build
-      - run: yarn test
-
-      - name: Build/release Docker images (Pro)
-        run: |
-          docker login -u $DOCKER_USER -p $DOCKER_PASSWORD
-          yarn build:docker:develop:pro
-          yarn release:docker:pro develop
-        env:
-          DOCKER_USER: ${{ secrets.DOCKER_USERNAME }}
-          DOCKER_PASSWORD: ${{ secrets.DOCKER_API_KEY }}
