--- .github/workflows/pro-release-selfhost.yml
@@ -1,41 +1,0 @@
-name: Budibase Release Selfhost (Pro)
-
-#  Temporary pipeline - eventualy this will be merged with the regular one
-
-on:
-  workflow_dispatch:
-
-jobs:
-  release:
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          node-version: 14.x
-          fetch_depth: 0
-
-      - name: Tag and release Docker images (Pro) (Self Host)
-        run: |
-          docker login -u $DOCKER_USER -p $DOCKER_PASSWORD
-
-          # Get latest release version
-          release_version=$(cat lerna.json | jq -r '.version')
-          echo "RELEASE_VERSION=$release_version" >> $GITHUB_ENV
-          release_tag=v$release_version
-
-          # Pull apps and worker images
-          docker pull budibase/apps-pro:$release_tag
-          docker pull budibase/worker-pro:$release_tag
-
-          # Tag apps and worker images
-          docker tag budibase/apps-pro:$release_tag budibase/apps-pro:$SELFHOST_TAG
-          docker tag budibase/worker-pro:$release_tag budibase/worker-pro:$SELFHOST_TAG
-
-          # Push images
-          docker push budibase/apps-pro:$SELFHOST_TAG
-          docker push budibase/worker-pro:$SELFHOST_TAG
-        env:
-          DOCKER_USER: ${{ secrets.DOCKER_USERNAME }}
-          DOCKER_PASSWORD: ${{ secrets.DOCKER_API_KEY }}
-          SELFHOST_TAG: latest
