+++ hosting/docker-compose.yaml
@@ -31,34 +31,6 @@ services:
       - worker-service
       - redis-service
 
-  app-service-pro:
-    restart: always
-    image: app-service-pro
-    container_name: bbapps-pro
-    ports:
-      - "${APP_PORT}:4002"
-    environment:
-      SELF_HOSTED: 1
-      COUCH_DB_URL: http://${COUCH_DB_USER}:${COUCH_DB_PASSWORD}@couchdb-service:5984
-      WORKER_URL: http://worker-service:4003
-      MINIO_URL: http://minio-service:9000
-      MINIO_ACCESS_KEY: ${MINIO_ACCESS_KEY}
-      MINIO_SECRET_KEY: ${MINIO_SECRET_KEY}
-      INTERNAL_API_KEY: ${INTERNAL_API_KEY}
-      BUDIBASE_ENVIRONMENT: ${BUDIBASE_ENVIRONMENT}
-      PORT: 4002
-      JWT_SECRET: ${JWT_SECRET}
-      LOG_LEVEL: info
-      SENTRY_DSN: https://a34ae347621946bf8acded18e5b7d4b8@o420233.ingest.sentry.io/5338131
-      ENABLE_ANALYTICS: "true"
-      REDIS_URL: redis-service:6379
-      REDIS_PASSWORD: ${REDIS_PASSWORD}
-    volumes:
-      - ./logs:/logs
-    depends_on:
-      - worker-service
-      - redis-service
-
   worker-service:
     restart: always
     image: budibase.docker.scarf.sh/budibase/worker
