+++ packages/server/.gitignore
@@ -8,5 +8,3 @@ db/dev.db/
 dist
 coverage/
 watchtower-hook.json
-pro.zip
-scripts/pro/*
